move lm head injection logic into lm_head.py and unify injection path for custom and hf impl
